requirements.txt
discord.py>=2.3.0
aiosqlite>=0.19.0
yt-dlp>=2023.10.13
PyNaCl>=1.5.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
google-generativeai>=0.3.0
asyncio-mqtt>=0.13.0
schedule>=1.2.0
pytz>=2023.3 